Extract non-configurable tests to JodaSpec
