+++ modules/joda/src/test/scala/pureconfig/module/joda/configurable/ConfigurableSuite.scala
@@ -1,12 +1,11 @@
 package pureconfig.module.joda.configurable
 
-import com.typesafe.config.{ ConfigFactory, ConfigValueFactory }
+import com.typesafe.config.ConfigFactory
 import org.joda.time._
-import org.joda.time.format.{ DateTimeFormat, DateTimeFormatter, ISOPeriodFormat, PeriodFormatter }
+import org.joda.time.format.{ DateTimeFormat, ISOPeriodFormat, PeriodFormatter }
 import org.scalacheck.{ Arbitrary, Gen }
 import pureconfig.syntax._
 import ConfigurableSuite._
-import pureconfig.module.joda._
 import pureconfig.configurable.{ ConfigurableSuite => JTime }
 import scala.collection.JavaConverters._
 
@@ -74,34 +73,6 @@ class ConfigurableSuite extends BaseSuite {
       conf.to[Conf].right.value shouldEqual Conf(yearMonth)
   }
 
-  it should "parse Instant" in forAll {
-    (instant: Instant) =>
-      val conf = ConfigFactory.parseString(s"""{instant:${instant.getMillis}}""")
-      case class Conf(instant: Instant)
-      conf.to[Conf].right.value shouldEqual Conf(instant)
-  }
-
-  it should "parse Interval" in forAll {
-    (interval: Interval) =>
-      val conf = ConfigFactory.parseString(s"""{interval:"${interval.toString}"}""")
-      case class Conf(interval: Interval)
-      conf.to[Conf].right.value shouldEqual Conf(interval)
-  }
-
-  it should "parse Duration" in forAll {
-    (duration: Duration) =>
-      val conf = ConfigFactory.parseString(s"""{duration:"${duration.toString}"}""")
-      case class Conf(duration: Duration)
-      conf.to[Conf].right.value shouldEqual Conf(duration)
-  }
-
-  it should "parse DateTimeZone" in forAll {
-    (dateTimeZone: DateTimeZone) =>
-      val conf = ConfigFactory.parseString(s"""{date-time-zone:"${dateTimeZone.getID}"}""")
-      case class Conf(dateTimeZone: DateTimeZone)
-      conf.to[Conf].right.value shouldEqual Conf(dateTimeZone)
-  }
-
   val periodFormatter: PeriodFormatter = ISOPeriodFormat.standard()
   implicit val periodInstance = periodConfigConvert(periodFormatter)
 
@@ -111,10 +82,6 @@ class ConfigurableSuite extends BaseSuite {
       case class Conf(period: Period)
       conf.to[Conf].right.value.period.toStandardDuration shouldEqual Conf(period).period.toStandardDuration
   }
-
-  checkRead[DateTimeFormatter](
-    DateTimeFormat.forPattern("yyyy-MM-dd'T'HH:mm:ss.SSSZZZ") ->
-      ConfigValueFactory.fromAnyRef("yyyy-MM-dd'T'HH:mm:ss.SSSZZZ"))
 }
 
 object ConfigurableSuite {
@@ -157,29 +124,9 @@ object ConfigurableSuite {
         zoneId <- zoneIdArbitrary.arbitrary
       } yield localDateTime.toDateTime(zoneId))
 
-  implicit val instantArbitrary: Arbitrary[Instant] =
-    Arbitrary(Arbitrary.arbitrary[Long].map(new Instant(_)))
-
-  implicit val intervalArbitrary: Arbitrary[Interval] =
-    Arbitrary(Arbitrary.arbitrary[(DateTime, DateTime)]
-      .filter(i => i._1.isBefore(i._2))
-      .map { i =>
-        // Comparing new Interval(foo) with Interval.parseWithOffset(foo) will return false
-        // See http://www.joda.org/joda-time/apidocs/org/joda/time/Interval.html#parseWithOffset-java.lang.String-
-        Interval.parseWithOffset(new Interval(i._1, i._2).toString)
-      })
-
-  implicit val durationArbitrary: Arbitrary[Duration] =
-    Arbitrary(Arbitrary.arbitrary[Long].map(new Duration(_)))
-
-  implicit val periodArbitrary: Arbitrary[Period] = {
-    Arbitrary(
-      for {
-        // Marshelling and unmarshalling fails with big values
-        i1 <- Gen.choose(-50, 50)
-        i2 <- Gen.choose(-50, 50)
-        i3 <- Gen.choose(-50, 50)
-        i4 <- Gen.choose(-50, 50)
-      } yield new Period(i1, i2, i3, i4))
-  }
+  implicit val periodArbitrary: Arbitrary[Period] =
+    Arbitrary(Arbitrary.arbitrary[Int].map { i =>
+      // Generated large longs crash Period constructor internally
+      new Period(i.toLong)
+    })
 }
